+++ pgtype/json.go
@@ -152,7 +152,7 @@ func (dst *JSON) Scan(src interface{}) error {
 func (src *JSON) Value() (driver.Value, error) {
 	switch src.Status {
 	case Present:
-		return string(src.Bytes), nil
+		return src.Bytes, nil
 	case Null:
 		return nil, nil
 	default: